Improved error handling Added root model to cache length and size Fixed bugs when errors occured Improved example application
Former-commit-id: c0f6217e837f4efb8ae75e6ff780111e60ffd478 [formerly 9b14bcd844cb77c8d1c3597df3cc429f95a3aa31]
Former-commit-id: 7f742bacbb6a72b91aab17ca137b39bea208d114

+++ lib/src/backend/impls/objectbox/models/src/store.dart
@@ -28,7 +28,7 @@ class ObjectBoxStore {
   final tiles = ToMany<ObjectBoxTile>();
 
   int length;
-  double size;
+  int size;
   int hits;
   int misses;
   String metadataJson;
